+++ CHANGELOG.md
@@ -123,6 +123,7 @@
 - foodsharers can't start the hygiene quiz !3898 @McGoldi
 - Hide Posts for non foodsaver users in Profile #2273 !3905 @SaiCode
 - Added strip_tags for renameConversation in MessageGateway #1111 !3907 @chriswalg
+- Fixed the redirect to the legal page #2206 !3839 @alex.simm
 
 ## Refactoring
 - Added computed column `is_sleeping` for `fs_foodsaver` #2198 !3437 !3809 @AntonBallmaier